fix: Enable the example projects to use locally linked versions of the Vime packages (#227)
* fix: preserve dir structure when bundling core icons

The local icons are used at development time, but were incorrectly copied without preserving the directory structure by cpy-cli. This commit moves that responsibility to a Stencil copy task.

* fix: ensure React example works with locally-linked package

Previously this would result in duplicate copies of React.

* fix: fix error in Rollup CSS build

* fix: make Stencil demo work when linked locally

* fix: Webpack example should always load extracted CSS

* fix: ensure Angular example works with linked packages

* build: add the examples as Lerna packages

This means that they can be used with locally linked versions of the @vime packages if desired. By default, the build command will only run on the public packages.

* build: fix Angular packages to v11-lts to ensure peer deps compatibility

* build: remove typescript devdependency from @vime/core

Stencil brings its own TypeScript with it, perhaps the version clash is breaking something?

* build: @vime/core - add @types/node, add DOM/ES2017 libs to tsconfig

* build: update @vime/core package-lock

* build: make prettier use its own ignore file

* chore: apply prettier formatting

* build: update validate workflow

* build: fix validate.yml format